Further changes to get element.height passed through
