+++ src/amp/components/elements/InteractiveAtomBlockComponent.tsx
@@ -11,46 +11,14 @@ import { ShowMoreButton } from '@root/src/amp/components/ShowMoreButton';
 // whitespace but give top-of-page atoms the chance to avoid resizing (or at
 // least show a good portion of the atom)
 
-const defaultInteractiveAtomHeight = 150;
-
-// Using the MobileHeight comment is ok (lets be very aware we're creating a
-// loosely-defined contract between interactive atoms and AMP here, that should
-// really be in CAPI)
-
-const heightRegex = /<!-- MobileHeight: (.*) -->/;
-
-//  The value of <!-- MobileHeight: value --> should be the height of an
-//  interactive atom at the widest mobile breakpoint to give a guesstimate to
-//  AMP for a reasonable height
-//
-// It should only be used on interactive atoms that are at the top of articles,
-// as it only matters when the bottom of the interactive atom is within the
-// viewport (otherwise AMP will auto-resize)
-//
-// Placeholder images must always exist for atoms that require using the
-// MobileHeight comment, otherwise the atom will not show on AMP
-//
-// Full page interactives should always have both
-
-const getHeight = (html?: string): number => {
-	if (html) {
-		const getHeightFromComment = heightRegex.exec(html);
-		if (
-			getHeightFromComment &&
-			typeof Number(getHeightFromComment[1]) === 'number'
-		) {
-			return Number(getHeightFromComment[1]); // Returns [ '<!-- MobileHeight: 100 -->', '100', index: 4, input: 'test<!-- MobileHeight: 100 -->', groups: undefined ]
-		}
-	}
-
-	return defaultInteractiveAtomHeight;
-};
+// const defaultInteractiveAtomHeight = 150;
 
 export const InteractiveAtomBlockComponent: React.FunctionComponent<{
 	url: string;
 	html?: string;
 	placeholderUrl?: string;
-}> = ({ url, placeholderUrl, html }) => {
+	height?: number;
+}> = ({ url, placeholderUrl, html, height }) => {
 	// On Dot Com, Interactive Atoms are sometimes used to modify the page
 	// around them. CSS, JS but no HTML can exist in an atom, and because we just
 	// add it to the page, it can modify the world around it.
@@ -59,11 +27,12 @@ export const InteractiveAtomBlockComponent: React.FunctionComponent<{
 	if (html === '') {
 		return null;
 	}
+	console.log('height in InteractiveAtomBlockComponent: ', height);
 
 	return (
 		<ClassNames>
 			{({ css }) => {
-				const styles = (height: number) => css`
+				const styles = css`
 					height: ${height}px;
 					width: 100%;
 					margin-top: 16px;
@@ -81,7 +50,7 @@ export const InteractiveAtomBlockComponent: React.FunctionComponent<{
 
 				return (
 					<amp-iframe
-						class={styles(getHeight(html))}
+						class={styles}
 						src={url}
 						layout="responsive"
 						sandbox="allow-scripts allow-same-origin allow-top-navigation-by-user-activation allow-popups"
